backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
celery[redis]>=5.3.0
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
passlib[bcrypt]>=1.7.4
bcrypt==4.2.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
requests>=2.31.0
pytest>=7.4.0
httpx>=0.25.0
pytest-cov>=4.1.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
